backend/requirements.txt
uvicorn
fastapi
pandas
pyyaml
numpy
ffmpeg-python  
pydantic[email]
motor   
cloudinary
python-dotenv